Add - Use heap allocation for http page buffer allowing large page sizes

+++ src/conf.h
@@ -191,7 +191,7 @@ typedef struct {
 	unsigned int fas_port;				//@brief Port the fas server will run on
 	int login_option_enabled;			//@brief Use default PreAuth Login script
 	int use_outdated_mhd;				//@brief Use outdated libmicrohttpd
-	int max_page_size;				//@brief Max page size to be served by libmicrohttpd
+	unsigned long long int max_page_size;		//@brief Max page size to be served by libmicrohttpd
 	int allow_preemptive_authentication;		//@brief Allow Preemptive Authentication using the ndsctl utility
 	int unescape_callback_enabled;			//@brief Enable external MHD unescape callback script
 	int fas_secure_enabled;			//@brief Enable Secure FAS
